requirements.txt
customtkinter==4.5.10
darkdetect==0.7.1
Pillow==9.2.0
pycodestyle==2.8.0
toml==0.10.2
